Handle wrappable token creation on coviction voting
